[AAE-10772] move LibraryMembershipDirective to content services
