+++ lib/content-services/src/lib/interfaces/index.ts
@@ -17,3 +17,4 @@
 
 export * from './node-allowable-operation-subject.interface';
 export * from './library-entity.interface';
+export * from './library-membership.interface';
