+++ lib/content-services/src/lib/interfaces/library-membership.interface.ts
@@ -15,15 +15,13 @@
  * limitations under the License.
  */
 
-export * from './highlight.directive';
-export * from './logout.directive';
-export * from './node-delete.directive';
-export * from './node-favorite.directive';
-export * from './node-restore.directive';
-export * from './node-download.directive';
-export * from './upload.directive';
-export * from './version-compatibility.directive';
-export * from './tooltip-card/tooltip-card.directive';
-export * from './infinite-select-scroll.directive';
+export interface LibraryMembershipToggleEvent {
+    updatedEntry?: any;
+    shouldReload: boolean;
+    i18nKey: string;
+}
 
-export * from './directive.module';
+export interface LibraryMembershipErrorEvent {
+    error: any;
+    i18nKey: string;
+}
