+++ lib/core/src/lib/breaking-changes.md
@@ -1,5 +1,6 @@
 6.0.0-beta.1
 - CheckAllowableOperationDirective: Moved from ADF Core to ADF content services
 - LibraryFavoriteDirective: Moved from ADF Core to ADF content services
+- LibraryMembershipDirective: Moved from ADF Core to ADF content services
